+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add UI for viewing page versions
 - Add feature to preview and restore to past versions
 - Add feature to add change summaries when saving or publishing
+- Improve library and script assignment UI by separating out global vs local
 
 ## 1.6.0